[gaim-migrate @ 3163] Please test this.
Yahoo changed it's authentication method and has started disallowing older clients (and third party clients using the older method) to connect to their servers.  This impliments the new authentication method so that we won't get blocked from Yahoo, and Indian users (who have already been blocked) can connect again.  It's not perfect--it should fail about 1 out of 16 times.  It's an easy bug to find and fix, but I want to go to bed now.  If you find yourself failing to connect more than this 1 of 16, please let me know, and I'll tell you how to help me debug it.

Thanks.