[gaim-migrate @ 5235] Aside from shifting lots of stuff around in aim.h, mostly just to annoy people because everything in the oscar directory will recompile, I took out a few calls to aimutil_getbleh and aimutil_putbleh in favor of other methods.  These functions should be used aaas little as possible because, um, well mid said so in his comments, I think.
The only real change is a fix for bug 710933, which says that
"Web Aware" is not set at signon, it is only set when you
explicitly set your status to anything.  Also, the I think the
interoperability flag changes how privacy/invisibility in ICQ
works, and I'm not really sure if it's good or not.  I think
it's more like AIM, but I'm too tired to figure out what that
means.
http://sourceforge.net/tracker/index.php?func=detail&aid=710993&group_id=235&atid=100235