[gaim-migrate @ 5500] this started out as simply adding an option to right-click on a jabber buddy and re-request authorization.  Then I ended up chasing the disgusting mess of const vs non-const parameters all over gaim.
The end result is that you can now right-click on jabber buddies and
re-request auth like you can for ICQ.  Also, a lot more things are const
that should be, I fixed a bug or two, and I cleaned up one of my least
favorite functions in gaim (linkify_text).  It is now decidedly less evil.