[gaim-migrate @ 6575] Fix a memleak in the status code (a temporary g_hash_table was not getting freed)
Make the auto-reconnect plugin fully multi-protocol aware.  Previously, if
it was waiting to sign on more than one account, and you unload the plugin,
it would continue to attempt to sign on all but one of the accounts.

Set wants_to_die to TRUE if Buddies->Signoff was chosen.

Get rid of some unneeded oscar debugging lines for faceprint :-)

Make oscar not attempt to signon again if you were disconnected because
you signed on from another location.