[gaim-migrate @ 7095] Bjoern Voigt writes: "As far as I can see, you commited a patch, which adds support for "Get User Info" on Yahoo and MSN. It's not my small Yahoo Get Info patch, it's the bigger one, which parses Yahoo and MSN profile pages.
I prepared a patch for it, which adds i18n macros to this patch.
Please commit the attached patch i18n21.patch.

The patch works fine for Yahoo contacts with English language
preference. Unfortunately the patch doesn't work with contacts with
other language preferences. The problem is, that in Yahoo every user
can set up a language for his/her profile. So
http://profiles.yahoo.com/CONTACT can deliver English, German, French,
... pages, depending on CONTACT's preference. The yahoo_get_info()
functions doesn't handle this and it's not easy to handle this.

I have some ideas, how to fix this issue. I like to discuss this with
the author of the patch or in gaim-devel."

committer: Luke Schierer <lschiere@pidgin.im>