[gaim-migrate @ 9230] Escape napster text.  I'm pretty sure Napster isn't supposed to support HTML tags.
Also, I can connect to a local napster server with no problems, but
when I try to connect to remote napster servers I get an error
when reading, and strerror(errno) says "Success."  Anyone know
what's up with that?  Maybe there isn't enough data in the TCP
stream or something and the socket is non-blocking, so it just
returns?