[gaim-migrate @ 9397] From out of nowhere comes a zephyr fix from yours truely.
If zephyr is the only account you sign on with, then the Gaim
login window never disappears because previously there was no
connection status dialog thingy.  This fixes that.

It would probably be good if this wasn't necessary.  I might
fix that a bit later.  Like, maybe make sure the login window
is hidden before the buddy list is drawn.