[gaim-migrate @ 10375] " Added the ability to remember your away message if you're disconnected and then reconnected." --Yosef Radchenko
Date: 2004-07-14 22:02
Sender: lschiere
Logged In: YES
user_id=28833

has this been tested with multiple accounts?


Date: 2004-07-14 22:49
Sender: jonrad
Logged In: YES
user_id=1083867

If you asking whether this was tested with multiple accounts
logged in at the same time, then yes. If you're asking
whether this was tested on multiple protocols, then no. I've
tested it on oscar. Also, I compiled it on my FreeBSD
machine and didn't get a chance to compile on any other
OSes, but conceptually it should work fine on others (But of
course, when dealing with computers, nothing works as it is
meant to).

Date: 2004-07-14 22:54
Sender: jonrad
Logged In: YES
user_id=1083867

Also, if you comment out lines 119 and 120:
if (gc->want_to_die)
   g_hash_table_remove(awayStates, aaccount);

Then it keeps the away information even if you purposesly
disconnected (As opposed to now, which only restores your
away state if you were kicked off the network or whatever)

Thats helpful if you want to test it.

committer: Luke Schierer <lschiere@pidgin.im>