[gaim-migrate @ 15262] Make UPnP truly asynchronous.
There are probably still a couple socket calls that should be made nonblocking, but I wanted to commit this before it became even bigger.  This contains a number of probable leak fixes in the upnp stuff.

The UPnP stuff has been updated to use gaim_url_fetch_request() instead of the specific implementation.

To make this all work, I had to make gaim_network_listen() and gaim_network_listen_range() also asynchronous - seems to work just fine apart from the UPnP calls seeming to take longer than they should (I'm planning to look into this).

I also triggered a STUN and UPnP discovery on startup so that we hopefully have the information when we need it.