[gaim-migrate @ 17779] I'm committing this to get it out of my tree...
As mentioned on gaim-devel, this normalizes some stuff.  Mark wants to standardize on "username" instead.  For recognition purposes, I'd prefer that, but using "screen name" is potentially useful in that users might not enter their OS username.  In any case, it's not a huge deal to me either way.  I believe the consensus before was for "screen name", and barring a consensus to change, "screen name" is the status quo anyway, so I'm making things consistent.