Make msim_remove_buddy() functional. It now sends the delbuddy, persist, and blocklist commands, each using msim_postprocess_outgoing().
Tested by removing a buddy, restarting Pidgin, and ensuring it did not reappear.

msim_postprocess_outgoing() and msim_postprocess_outgoing_cb() now both use msim_do_postprocessing() to do the actual postprocessing -- insertion of the
uid into the message. msim_postprocess_outgoing() calls msim_do_postprocessing()
if the uid is immediately available (for example, from the buddy list), while
msim_postprocess_outgoing_cb() calls it when the uid was looked up from the
server.

The new function avoids code duplication, which is important now because of the
additional feature of having "<uid>" within a string replaced with the uid, i
instead of adding a new integer field of the uid. msim_remove_buddy uses
this to insert the uid into the "body" field.

--HG--
branch : soc.2007.msimprpl