Only say we are connected after the MsimSession structure is fully setup. (Also change a few debugging statements to help find this bug.)
Previously, purple_connection_set_state(..., PURPLE_CONNECTED) was called
too early, so that if a client called any msimprpl functions that operated
on MsimSession (pretty much all of them) within the signed-on signal handler,
the userid etc., would not yet be correct and the call would fail. Now, the
call is called at the right time so it works as expected.

Closes #2520.

--HG--
branch : soc.2007.msimprpl