disapproval of revision '9ab66eb1cd22e33bb9ba8d5229ed8ff20b39ed95'
This really needs to happen with the appropriate magic necessary to allow
adding newer functions when necessary without forcing a bump in the
major. I am not necessarily succumbing to 'fear of incrementing a version
number' or anything, I just don't think this particular feature is
important or useful enough. I would be interested to see if there's still
any confusion left after the change from 'Screenname' to 'Username'.

--HG--
branch : next.minor