Pluck the MSN bug fixes.
*** Plucked rev e3608f6f (qulogic@pidgin.im):
Ignore the Forward List status bit when checking if user is on the Allow
and Block lists at the same time.

References #6702.

*** Plucked rev 99d00e33 (qulogic@pidgin.im):
For federated buddies, we should set <Member>/<Type> to Email, and use an
<Email> node instead of <PassportName>. Along with that, we shouldn't set
any <Annotations> to specify it's a federated buddy either, since the MSN
server will complain about it.

Fixes #7939.

*** Plucked rev bf71505b (qulogic@pidgin.im):
Fix a faulty condition that would result in an attempt to add most of your
email-only contacts to your buddy and/or privacy list. This was temporary,
but resulted in a lot of extra traffic during login.

References #8579.

--HG--
branch : release-2.5.6