disapproval of revision '7da5ce54f293e74640a719de8b7ae086846536fd'
Conversation in devel@c.p.i leads to deryni saying:
Ok, if we send charset=utf-8 then we need to have sent a utf8 username and realm in digest-response but need to have used iso8859-1 versions of them when calculating response-value (if they are valid iso8859-1). I think.

which is what the code was doing before.