*** Plucked rev d0e53002eb30c6dbf9735ca50784d42141dfc96a (sulabh@soc.pidgin.im): Yahoo version 16 login. ---yet to do: prevent password from being displayed in debug logs---
*** Plucked rev 398170eb3baade1ddeeca492dbb74d023c6d8a27 (sulabh@soc.pidgin.im):
Inform user that he wont be able to use Yahoo! in case of absence of ssl
support

*** Backported rev a6ccff1d5368c712499f9b2f3671fa017f8efca2 (rekkanoryo@pidgin.im):
Support logging into Yahoo! Japan with protocol 16.  It seems not to care
what version bytes are sent once you use the right URL's to log in.

Also changelogged the fact that we have support for Yahoo 16 login.

--HG--
branch : release-2.5.7