applied changes from 76f0c314a1c503b99d7343d958c0565af361cfff              through 13c4f11294b803865ced5e10b4d498540a31e3df
Plucked this revision from im.pidgin.pidgin
Original commit message:
Fix a possible crash when receiving URLs in yahoo.  It doesn't crash for me
but it did for the guy reporting the bug.  Might be architecture
dependent or compile option dependent.  In any case this is bad and we
should release 2.6.1 tonight or tomorrow

--HG--
branch : release-2.6.1