Fix some problems adding "stuff" to oscar buddy lists, where "stuff" is often buddies to block, but can also be buddies to allow or maybe your privacy setting.  I imagine this eliminates some "unable to add buddy" errors.
Apparently I discovered this idiosyncrasy on 2007-04-23 and fixed it for
icon items in 39b38bcf07260d6d73a20ae949b92fcb51405c8c, but for some
reason I didn't fix it for any other types of items.

I'm also making this loop a bit less heavy handed... it used to require
that new items had an itemID greater than every other item.  Now it just
makes sure the itemID is not equal the itemID and not equal to the groupID
of any other item.

Let it be known that the best way for me to fix blocking is to have people
I don't want to talk to IM me.