*** Plucked rev 4be2df4f72bd8a55cdae7f2554b73342a497c92f (bcc0147bab874ca52c55ad4900545e17528bf8fd): There's no need to check type twice here. Also, g_return_if_reached is not correct since img==NULL is caused by invalid network input and not some programmer error.
*** Plucked rev 3d02401cf232459fc80c0837d31e05fae7ae5467 (bcc0147bab874ca52c55ad4900545e17528bf8fd):
Using 'if (img==NULL) return;' fails to clean up properly, so invert the
logic of that section.

*** Plucked rev c64a1adc8bda2b4aeaae1f273541afbc4f71b810 (bcc0147bab874ca52c55ad4900545e17528bf8fd):
Each PurpleSmiley has a PurpleStoredImage behind it, so there's no need to
go and create a new one when someone asks for one.

Also, ignore requests for smileys that don't exist in the image store.

--HG--
branch : release-2.6.5