Apply Mark's specific changes listed below to the 2.7.7 branch.
*** Plucked rev 05ca7c0f8c782c05e9866a9ac6ccc03acc7d2c7f (markdoliner@pidgin.im):
Close open requests related to this xfer when the request is canceled
locally.  For oscar this includes disconnecting when you have an
incoming transfer request.  Without this change Pidgin will crash if
the user tries to interact with the dialog.  This change fixes #11666.

Now instead of crashing we'll leak.  See the lengthy comment in the
code if anyone wants to fix this.

*** Plucked rev 089c261f1de00667abd623ce3c5b471e91b09016 (markdoliner@pidgin.im):
I noticed a NULL printf crash from the first chunk of this change.
In the second chunk I changed the code to match the error message from
the first chunk.  I prefer this message.

*** Plucked rev 6e8da78b6e5ccdafa85c8afebff37e426d9a58d3 (markdoliner@pidgin.im):
Move the call to flap_connection_schedule_destroy from oscar_chat_kill
to oscar_chat_leave.  This avoids having flap_connection_schedule_destroy
called from purple_connerr, which itself is called by flap_connection_destroy_cb
I'm hoping this change fixes #5927, the oscar crash when a flap connection
is disconnected.

--HG--
branch : release-2.7.7