Hopefully fixes #13298, the bug about not being able to add MSN buddies. Here's what I think is happening: 1. Before we add the buddy we issue an FQY request to ask the server what    network the username is for (either normal MSN or Yahoo) 2. When we get the response we add the buddy to our buddy list with the    network ID given to us by the server 3. For some reason the server is now returning a network ID of 0 ("unknown")    instead of 1 ("normal MSN") for normal passport buddies, and we bail out    when we encounter this.  QuLogic thinks the server used to return 1 in    this case.
My change is to just not bail out if the FQY response has network ID 0.
Instead of treat 0 as 1 and continue with the add.  It looks like our
SOAP request to add the buddy to our address book will fail a big further
down the road if the buddy doesn't exist--so we're still protecting
against that.