*** Plucked rev 6edd4aa8f524b904f6d0c0e4462ebffb9e6f98d1 (markdoliner@pidgin.im): Change the heuristic we use for setting od->icq.  We used to set it to true any time the username contained only digits.  Now we set it to try if account->protocol_id is prpl-icq.  This means we're now relying on the user to tell us whether their account is AIM or ICQ... and I think that's fine.
The reason for this change is that we/AOL apparently allow the user to
login to ICQ accounts by entering their email address.  This means we're
no longer able to look at the username to determine whether it's an AIM
name or an ICQ name.

This was a problem because we were trying to format the spacing/capitalization
in the username if the name was an icq email address, which resulted in
an annoying popup error at login.  Fixes #13883.

It's probably safe for this to go into im.pidgin.pidgin (maybe after the
pending release), but I really don't want to keep changing the 2.x.y branch.
I think we need to move on to 3.x.y.

--HG--
branch : release-2.x.y