toc: Remove duplicate calls to purple_normalize
CodeQL thinks there's a "potentially overflowing call to snprintf", but
it doesn't realize that `purple_normalize` should return the same thing
for the two calls. But it's safer to not rely on that by avoiding the
duplicate calls.

Unfortunately, this doesn't fix the CodeQL result, but it's easier to
reason about directly that it's a false positive.