Fix i18n support
None of the plugins were binding their gettext domain so the translations
would never work.

Napster was still setup for the plugin pack which obviously didn't work, but
we didn't notice this because we didn't add an option for NLS which made the
code not get compiled.

All of the glib gettext helpers that we use were added before or in 2.18 and
we require 2.18 on windows so we should be fine in that reguard.