The ctype(3) isspace, isdigit, ... functions take an int argument that must be in the range [0-255] or EOF. Passing it a signed char results in undefined behaviour.
Some implementations of libc, such as glibc as of 2018, attempt to avoid
the worst of the undefined behavior by defining the functions to work for
all integer inputs representable by either unsigned char or char.

On NetBSD, ctype(3) functions will crash with a SIGSEGV signal on invalid inputs
as a diagnostic aid for applications.

This set of patches ensures the char is first converted to unsigned char to ensure
that the values are within the correct range.